updating fork

+++ Assets/Scripts/Character/IDamageable.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public interface IDamageable
 {
 
-    void Damage(int damage);
+    void Damage(float damage);
 
-    void Heal(int healing);
+    void Heal(float healing);
 
     void Kill();
